refactor: 컴포넌트 이름, 스토리 수정
